worker: Fix the CentOS 7 container
Remove yum check-update as a newer version of epel-release package is
available in the epel repository and the commands fail.

The versions of sip and python-qt5 used in the container are no longer
available in the Fedora repositories so build them using fedpkg command
from a specific commits in order to get the same versions as before.

Change the value of python3_pkgversion to 36 as otherwise some
dependencies can't be installed as they are only available as python36-*
package.